third_party/blink/web_tests/external/wpt/web-animations/responsive/columnGap-expected.txt

This is a testharness.js-based test.
[PASS] column-gap responsive to style changes
[FAIL] column-gap clamped to 0px on keyframes
  assert_equals: expected "20px" but got "0px"
[PASS] column-gap responsive to inherited changes
Harness: the test ran to completion.
